+++ sunpy/net/hek2vso/hek2vso.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 # Author:   Michael Malocha <mjm159@humboldt.edu>
-# Last Edit:  August 10th, 2013
 #
 # This module was developed with funding from the GSOC 2013 summer of code
 #
@@ -20,8 +19,8 @@ from sunpy.net import hek
 from sunpy.net import vso
 from sunpy.util.progressbar import TTYProgressBar
 
-__author__ = 'Michael Malocha'
-__version__ = 'Aug 10th, 2013'
+__author__ = ['Michael Malocha',]
+__emails__ = ['<mjm159@humboldt.edu>',]
 
 __all__ = ['translate_results_to_query', 'vso_attribute_parse', 'H2VClient']
 
@@ -169,7 +168,7 @@ class H2VClient(object):
         hek_results: sunpy.net.hek.hek.Response or list of Responses
             The results from a HEK query in the form of a list.
         limit: int
-            An approximate limit to the desired number of VSO results.
+            An approximate limit to the desired number of VSO result records.
         progress: Boolean
             A flag to turn off the progress bar, defaults to "off"