+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Latest
  * CD matrix now updated, if present, when Map pixel size is changed.
  * Removed now-redundant method for rotating IRIS maps since the functionality exists in Map.rotate()
  * Provide `__str__` and `__repr__` methods on vso `QueryResponse` deprecate `.show()`
+* Added tests for the sunpy.net.hek2vso module
  * SunPy colormaps are now registered with matplotlib on import of `sunpy.cm`
  * `sunpy.cm.get_cmap` no longer defaults to 'sdoaia94'
  * Added database url config setting to be setup by default as a sqlite database in the sunpy working directory